requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.28.0=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.3.0=pypi_0
annotated-types=0.6.0=pypi_0
anthropic=0.25.6=pypi_0
anyio=4.3.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
blinker=1.7.0=pypi_0
blis=0.7.11=pypi_0
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.3=pypi_0
catalogue=2.0.10=pypi_0
certifi=2024.2.2=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloud-sql-python-connector=1.9.1=pypi_0
cloudpathlib=0.16.0=pypi_0
confection=0.1.4=pypi_0
cryptography=42.0.5=pypi_0
cuda-cudart=12.4.127=0
cupy-cuda11x=12.3.0=pypi_0
cymem=2.0.8=pypi_0
cython=0.29.37=pypi_0
dataclasses-json=0.6.4=pypi_0
deprecated=1.2.14=pypi_0
dirtyjson=1.0.8=pypi_0
distro=1.9.0=pypi_0
en-core-web-sm=3.7.1=pypi_0
exceptiongroup=1.2.0=pypi_0
faiss-gpu=1.8.0=py3.9_h4c7d538_0_cuda12.1.1
fastrlock=0.8.2=pypi_0
feedparser=6.0.8=pypi_0
filelock=3.13.3=pypi_0
flask=2.0.2=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
ftfy=6.2.0=pypi_0
gensim=4.3.2=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-ai-generativelanguage=0.6.2=pypi_0
google-api-core=2.18.0=pypi_0
google-api-python-client=2.127.0=pypi_0
google-auth=2.29.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-generativeai=0.5.2=pypi_0
googleapis-common-protos=1.63.0=pypi_0
greenlet=3.0.3=pypi_0
grpcio=1.62.2=pypi_0
grpcio-status=1.62.2=pypi_0
h11=0.14.0=pypi_0
hdbscan=0.8.33=pypi_0
httpcore=1.0.5=pypi_0
httplib2=0.22.0=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.20.3=pypi_0
idna=3.6=pypi_0
importlib-metadata=7.1.0=pypi_0
inflect=7.2.1=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
itsdangerous=2.2.0=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
langchain=0.1.13=pypi_0
langchain-community=0.0.29=pypi_0
langchain-core=0.1.36=pypi_0
langchain-text-splitters=0.0.1=pypi_0
langcodes=3.4.0=pypi_0
langsmith=0.1.38=pypi_0
language-data=1.2.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libcublas=12.4.5.8=0
libfaiss=1.8.0=h046e95b_0_cuda12.1.1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
llama-index=0.10.25=pypi_0
llama-index-agent-openai=0.2.1=pypi_0
llama-index-cli=0.1.11=pypi_0
llama-index-core=0.10.25.post3=pypi_0
llama-index-embeddings-huggingface=0.2.0=pypi_0
llama-index-embeddings-openai=0.1.7=pypi_0
llama-index-indices-managed-llama-cloud=0.1.5=pypi_0
llama-index-legacy=0.9.48=pypi_0
llama-index-llms-anyscale=0.1.3=pypi_0
llama-index-llms-huggingface=0.1.4=pypi_0
llama-index-llms-langchain=0.1.3=pypi_0
llama-index-llms-ollama=0.1.2=pypi_0
llama-index-llms-openai=0.1.14=pypi_0
llama-index-multi-modal-llms-openai=0.1.4=pypi_0
llama-index-program-openai=0.1.5=pypi_0
llama-index-question-gen-openai=0.1.3=pypi_0
llama-index-readers-file=0.1.13=pypi_0
llama-index-readers-llama-parse=0.1.4=pypi_0
llama-parse=0.4.0=pypi_0
llamaindex-py-client=0.1.15=pypi_0
locate=1.1.1=pypi_0
marisa-trie=1.1.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
marshmallow=3.21.1=pypi_0
mdurl=0.1.2=pypi_0
minijinja=1.0.16=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
more-itertools=10.2.0=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.8=pypi_0
multidict=6.0.5=pypi_0
murmurhash=1.0.10=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.2.1=pypi_0
nltk=3.8.1=pypi_0
numpy=1.22.4=pypi_0
numpy-base=1.26.4=py39hb5e798b_0
nvidia-cublas-cu11=11.11.3.6=pypi_0
nvidia-cuda-cupti-cu11=11.8.87=pypi_0
nvidia-cuda-nvrtc-cu11=11.8.89=pypi_0
nvidia-cuda-runtime-cu11=11.8.89=pypi_0
nvidia-cudnn-cu11=8.7.0.84=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.3.0.86=pypi_0
nvidia-cusolver-cu11=11.4.1.48=pypi_0
nvidia-cusparse-cu11=11.7.5.86=pypi_0
nvidia-nccl-cu11=2.19.3=pypi_0
nvidia-nvtx-cu11=11.8.86=pypi_0
openai=1.23.6=pypi_0
openssl=3.0.13=h7f8727e_0
orjson=3.10.0=pypi_0
packaging=23.2=py39h06a4308_0
pandas=2.2.1=pypi_0
pillow=10.2.0=pypi_0
pip=24.0=pypi_0
preshed=3.0.9=pypi_0
proto-plus=1.23.0=pypi_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=pypi_0
pyarrow=16.0.0=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycparser=2.22=pypi_0
pydantic=2.6.4=pypi_0
pydantic-core=2.16.3=pypi_0
pydeck=0.9.0b1=pypi_0
pygments=2.17.2=pypi_0
pymupdf=1.24.0=pypi_0
pymupdfb=1.24.0=pypi_0
pymysql=1.1.0=pypi_0
pyparsing=3.1.2=pypi_0
pypdf=4.1.0=pypi_0
python=3.9.19=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.0=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.0=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.2=pypi_0
scikit-learn=1.0.1=pypi_0
scipy=1.10.1=pypi_0
sentence-transformers=2.6.1=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=69.5.1=pypi_0
sgmllib3k=1.0.0=pypi_0
six=1.16.0=pypi_0
smart-open=6.4.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
spacy=3.7.4=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlalchemy=2.0.29=pypi_0
sqlite=3.41.2=h5eee18b_0
sqlitedict=1.7.0=pypi_0
srsly=2.4.8=pypi_0
streamlit=1.33.0=pypi_0
striprtf=0.0.26=pypi_0
sympy=1.12=pypi_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.2.3=pypi_0
thinc=8.2.3=pypi_0
threadpoolctl=3.4.0=pypi_0
tiktoken=0.6.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.2=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.1=pypi_0
torch=2.2.2+cu118=pypi_0
torchaudio=2.2.2+cu118=pypi_0
torchvision=0.17.2+cu118=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.2=pypi_0
transformers=4.39.2=pypi_0
triton=2.2.0=pypi_0
typeguard=4.2.1=pypi_0
typer=0.9.4=pypi_0
typing-extensions=4.10.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.1=pypi_0
uritemplate=4.1.1=pypi_0
urllib3=2.2.1=pypi_0
wasabi=1.1.2=pypi_0
watchdog=4.0.0=pypi_0
wcwidth=0.2.13=pypi_0
weasel=0.3.4=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.43.0=pypi_0
wordfreq=3.1.1=pypi_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_0
yarl=1.9.4=pypi_0
zipp=3.18.1=pypi_0
zlib=1.2.13=h5eee18b_0
